% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{write_git_ignore}
\alias{write_git_ignore}
\title{Write .gitignore file}
\usage{
write_git_ignore()
}
\description{
This includes the following line items:
.Rproj.user
.Rhistory
.RData
.Ruserdata
.xlsx
.pdf
.html
/data
/output
}
\details{
The main purpose is to ensure no private data gets posted
}
