% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{write_index}
\alias{write_index}
\title{Write Index}
\usage{
write_index()
}
\description{
This is the main file for the project which will be used to source
all functions, libraries, and data as necessary and run all other
scripts.
}
